Creacion de las clases ItemMenu,Bebida,Plato y Categoria

+++ src/main/java/nc/diedtp/Bebida.java
@@ -27,7 +27,7 @@ public class Bebida extends ItemMenu{
         return peso;
         }
 
-    public int getGraduacionAlcoholica() {
+    public float getGraduacionAlcoholica() {
         return graduacionAlcoholica;
     }
